refactor(admin-dogs-diseases): split disease browser internals
  Extract private DB helpers and web rendering helpers into
  feature-local internal modules to shrink the main repository
  and page client files. Behavior, API shape, URL state,
  filtering, pagination, and rendered output stay unchanged.

  Files:
  - apps/web/components/admin/dogs/diseases/admin-dog-diseases-
    page-client.tsx
  - apps/web/components/admin/dogs/diseases/internal/disease-
    formatters.ts
  - apps/web/components/admin/dogs/diseases/internal/disease-
    results.tsx
  - packages/db/admin/dogs/diseases/internal/disease-
    selection.ts
  - packages/db/admin/dogs/diseases/internal/parent-preview.ts
  - packages/db/admin/dogs/diseases/list-dog-diseases.ts
